Tue, Jun 18, 2024, 10:11 PM +05:30

+++ Lecture-1-C/Practice Set/ps3.c
@@ -5,6 +5,6 @@
 int main(void) {
     int num = get_int("Guess a number between 1 to 100: ");
     printf("%i\n", num);
-    long rand = random();
-    printf("Random: %lu\n", rand);
+    float rand = random();
+    printf("Random: %f\n", rand);
 }
